Relocate state folder
